+++ common/ios/doomengine/iphone_async.cpp
@@ -360,8 +360,8 @@ void SendSetupPacketIfNecessary() {
 		if ( setupPacket.playerID[i] == 0 ) {
 			continue;
 		}
-		int r = sendto( gameSocket, &setupPacket, sizeof( setupPacket ), 0, 
-					   &netPlayers[i].peer.address, sizeof( netPlayers[i].peer.address ) );
+        ssize_t r = sendto( gameSocket, &setupPacket, sizeof( setupPacket ), 0,
+                           &netPlayers[i].peer.address, sizeof( netPlayers[i].peer.address ) );
 		if ( r == -1 ) {
 			Com_Printf( "UDP sendTo failed: %s\n", strerror( errno ) );
 			close( gameSocket );
@@ -919,7 +919,7 @@ void iphoneAsyncTic() {
 							}
 						}
 					}
-					int	packetSize = (byte *)cmd_p - (byte *)&gp;
+					long packetSize = (byte *)cmd_p - (byte *)&gp;
 					(void)packetSize;
 					
 					// use the most recent tic that both the client and
